Clubbed Employee Role, Status and Task Type into Enum Rest Controller
Since, all the Controllers were providing similar responses, therefore clubbed into single response controller.
